refactor(stories): sort sections by what each story is really for
Four stories were filed by the folder they arrived in rather than by what
they teach, and one section was nested for no reason its content
justified.

- `Examples/Forces/Clustering` was the only title with another title
  beneath it, which made Forces a sidebar folder instead of a component —
  one section reading as a different kind of thing from the other eight.
  Sections are flat again and `storySort` drops the nested entry.
- `with-labels` becomes Forces/"Clusters with Labels": it calls
  `setPointClusters` itself and reads `getClusterPositions` for the
  centroid labels, so it is the real cluster reference.
- `radial` and `worm` move to Showcase. Radial hands `generateMeshData`
  straight to `createCosmos`, so every cluster call is in the helper and
  the file teaches nothing; Worm keeps its loop inline but is passive eye
  candy, and presentation is what Showcase sorts on. Showcase is now four
  stories that all build on `generateMeshData` and exist for the look of
  them — a sharper rule than it had as the former Misc.
- `actions` moves to Get Started, which now reads Quick Start → Actions:
  build a graph, then see what you can do with it. Tagged `beginner` so
  the tag filter surfaces the whole on-ramp, not just its first step.
- `position-rescaling` moves to Points. `rescalePositions` governs how
  the coordinates passed to `setPointPositions` map into space, which is
  a positions concern, so it leads the section ahead of appearance.

Interaction is left with four stories that are all about input. The two
in-repo links to Actions are repointed, and every Storybook deep link is
checked against the generated index.
