+++ src/stories/get-started/actions/style.css
