+++ src/stories/points.stories.ts
@@ -5,20 +5,35 @@ import { CosmosStoryProps } from './create-cosmos'
 import { allShapes } from './points/shapes'
 import { imageExample } from './points/images'
 import { pointLabels } from './points/labels'
-import { worm } from './points/animated-colors'
+import { moscowMetroStations } from './points/position-rescaling'
 
-import createCosmosRaw from './create-cosmos?raw'
-import generateMeshDataRaw from './generate-mesh-data?raw'
 import shapesStoryRaw from './points/shapes/index?raw'
+import rescalingStoryRaw from './points/position-rescaling/index?raw'
+import rescalingCoordsRaw from './points/position-rescaling/moscow-metro-coords?raw'
+import rescalingColorsRaw from './points/position-rescaling/point-colors?raw'
+import rescalingCssRaw from './points/position-rescaling/style.css?raw'
 import imagesStoryRaw from './points/images/index?raw'
 import labelsStoryRaw from './points/labels/index?raw'
 import labelsDataRaw from './points/labels/data?raw'
 import labelsLabelsRaw from './points/labels/labels?raw'
 import labelsCssRaw from './points/labels/style.css?raw'
-import animatedColorsRaw from './points/animated-colors?raw'
 
 const meta: Meta<CosmosStoryProps> = {
   title: 'Examples/Points',
+}
+
+export const PositionRescaling: Story = {
+  ...createStory(moscowMetroStations),
+  name: 'Position Rescaling',
+  tags: ['beginner', 'interactive'],
+  parameters: {
+    sourceCode: [
+      { name: 'Story', code: rescalingStoryRaw },
+      { name: 'moscow-metro-coords', code: rescalingCoordsRaw },
+      { name: 'point-colors', code: rescalingColorsRaw },
+      { name: 'style.css', code: rescalingCssRaw },
+    ],
+  },
 }
 
 export const AllShapes: Story = {
@@ -92,19 +107,6 @@ export const PointLabels: Story = {
       { name: 'data.ts', code: labelsDataRaw },
       { name: 'labels.ts', code: labelsLabelsRaw },
       { name: 'style.css', code: labelsCssRaw },
-    ],
-  },
-}
-
-export const AnimatedColors: Story = {
-  ...createStory(worm),
-  name: 'Animating Point Colors',
-  tags: ['large-data'],
-  parameters: {
-    sourceCode: [
-      { name: 'Story', code: animatedColorsRaw },
-      { name: 'create-cosmos', code: createCosmosRaw },
-      { name: 'generate-mesh-data', code: generateMeshDataRaw },
     ],
   },
 }
