+++ src/stories/points/position-rescaling/index.ts
