+++ src/stories/points/position-rescaling/moscow-metro-coords.ts
