+++ src/stories/points/position-rescaling/point-colors.ts
