+++ src/stories/points/position-rescaling/style.css
