+++ src/stories/welcome.mdx
@@ -72,7 +72,7 @@ graph.render()
 Try the [Quick Start](../?path=/story/examples-get-started--quick-start) example.
 
 ### Examples
-- [Actions](../?path=/story/examples-interaction--actions)
+- [Actions](../?path=/story/examples-get-started--actions)
 - [Adding Point Labels](../?path=/story/examples-points--point-labels) (via [`@interacta/css-labels`](https://www.npmjs.com/package/@interacta/css-labels))
